[redistribute] work in progress

+++ maia/utils/parallel/utils.py
@@ -9,8 +9,8 @@ from maia       import npy_pdm_gnum_dtype
 def gathering_distribution(n_elt, comm):
   """
   """
-  if comm.Get_rank() == 0: distrib = np.array([0    , n_elt, n_elt ], dtype=np.int32)
-  else                   : distrib = np.array([n_elt, n_elt, n_elt ], dtype=np.int32)
+  if comm.Get_rank() == 0: distrib = np.array([0    , n_elt, n_elt ], dtype=npy_pdm_gnum_dtype)
+  else                   : distrib = np.array([n_elt, n_elt, n_elt ], dtype=npy_pdm_gnum_dtype)
   return distrib
 
 def uniform_distribution(n_elt, comm):
